chore: Start moving namespace-select-filter to own folder
