+++ packages/core/src/renderer/components/namespace-select-filter/namespace-select-filter.scss
