+++ packages/core/src/renderer/components/namespace-select-filter/namespace-select-filter.test.tsx
@@ -24,8 +24,8 @@ import type { Disposer } from "@k8slens/utilities";
 import { disposer } from "@k8slens/utilities";
 import { renderFor } from "../test-utils/renderFor";
 import { NamespaceSelectFilter } from "./namespace-select-filter";
-import type { NamespaceStore } from "./store";
-import namespaceStoreInjectable from "./store.injectable";
+import type { NamespaceStore } from "../namespaces/store";
+import namespaceStoreInjectable from "../namespaces/store.injectable";
 
 function createNamespace(name: string): Namespace {
   return new Namespace({
