+++ packages/core/src/renderer/components/namespace-select-filter/namespace-select-filter.tsx
@@ -8,9 +8,9 @@ import "./namespace-select-filter.scss";
 import React, { useEffect, useRef } from "react";
 import { observer } from "mobx-react";
 import { withInjectables } from "@ogre-tools/injectable-react";
-import type { NamespaceSelectFilterModel, NamespaceSelectFilterOption } from "./namespace-select-filter-model/namespace-select-filter-model";
-import { selectAllNamespaces } from "./namespace-select-filter-model/namespace-select-filter-model";
-import namespaceSelectFilterModelInjectable from "./namespace-select-filter-model/namespace-select-filter-model.injectable";
+import type { NamespaceSelectFilterModel, NamespaceSelectFilterOption } from "./namespace-select-filter-model";
+import { selectAllNamespaces } from "./namespace-select-filter-model";
+import namespaceSelectFilterModelInjectable from "./namespace-select-filter-model.injectable";
 import { VariableSizeList } from "react-window";
 import { Icon } from "../icon";
 import { cssNames, prevDefault } from "@k8slens/utilities";
