+++ packages/core/src/renderer/components/namespaces/filter-by-namespace.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import namespaceStoreInjectable from "../store.injectable";
+import namespaceStoreInjectable from "./store.injectable";
 
 export type FilterByNamespace = (namespace: string) => void;
